+++ src/lib.rs
@@ -85,7 +85,7 @@ gmp-mpfr-sys in your crate, add it as a dependency inside
 
 ```toml
 [dependencies]
-gmp-mpfr-sys = "1.2"
+gmp-mpfr-sys = "1.3"
 ```
 
 This crate required rustc version 1.37.0 or later.
@@ -99,7 +99,7 @@ Rust bindings but do not use the C libraries directly.
 If on the other hand a dependent crate makes use of internal
 implementation details, or includes a C library that directly uses the
 header (*.h*) and library (*.a*) files built using C, it can be a good
-idea to depend on version `"~1.2"` instead of version `"1.2"` in order
+idea to depend on version `"~1.3"` instead of version `"1.3"` in order
 to ensure backwards compatibility at the C level as well.
 
 ## Optional features
@@ -118,7 +118,7 @@ selectively, you can add the dependency like this to [*Cargo.toml*]:
 
 ```toml
 [dependencies.gmp-mpfr-sys]
-version = "1.2"
+version = "1.3"
 default-features = false
 features = ["mpfr"]
 ```
@@ -276,7 +276,7 @@ To use a different directory, you can set the environment variable
 */
 #![no_std]
 #![warn(missing_docs)]
-#![doc(html_root_url = "https://docs.rs/gmp-mpfr-sys/~1.2")]
+#![doc(html_root_url = "https://docs.rs/gmp-mpfr-sys/~1.3")]
 #![doc(html_logo_url = "https://tspiteri.gitlab.io/gmp-mpfr-sys/rug.svg")]
 #![doc(test(attr(deny(warnings))))]
 #![cfg_attr(feature = "fail-on-warnings", deny(warnings))]