+++ RELEASES.md
@@ -10,15 +10,17 @@ Version 1.3.0 (unreleased)
 
   * The implementation details of
     <code>[gmp][gmp-1-3]::[mpz\_t][gmp-mpz-1-3]</code>,
-    <code>[gmp][gmp-1-3]::[mpf\_t][gmp-mpf-1-3]</code> and
-    <code>[gmp][gmp-1-3]::[randseed\_t][gmp-randseed-1-3]</code>
-	have been
+    <code>[gmp][gmp-1-3]::[mpf\_t][gmp-mpf-1-3]</code>,
+    <code>[gmp][gmp-1-3]::[randseed\_t][gmp-randseed-1-3]</code> and
+    <code>[mpfr][mpfr-1-3]::[mpfr\_t][mpfr-mpfr-1-3]</code> have been
     changed to use [`NonNull`] instead of [mutable pointers][pointer].
 
 [gmp-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/gmp/index.html
 [gmp-mpf-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/gmp/struct.mpf_t.html
 [gmp-mpz-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/gmp/struct.mpz_t.html
 [gmp-randseed-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/gmp/struct.randseed_t.html
+[mpfr-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/mpfr/index.html
+[mpfr-mpfr-1-3]: https://docs.rs/gmp-mpfr-sys/~1.3/gmp_mpfr_sys/mpfr/struct.mpfr_t.html
 
 Version 1.2.4 (2020-06-14)
 ==========================