+++ RELEASES.md
@@ -5,7 +5,7 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
-Version 1.3.0 (unreleased)
+Version 1.3.0 (2020-07-13)
 ==========================
 
   * [MPFR] was updated from version 4.0.2-p9 to 4.1.0-rc2.